functions/get_files_info.py
import os
from google.genai import types


def get_files_info(working_directory, directory="."):
    try:
        working_dir_abs = os.path.abspath(working_directory)

        # if not os.path.isdir(directory):
        #     raise Exception(f'Error: "{directory}" is not a directory')

        target_dir = os.path.normpath(os.path.join(working_dir_abs, directory))

        if not os.path.isdir(target_dir):
            raise Exception(f'Error: "{directory}" is not a directory')

        # Will be True or False
        valid_target_dir = os.path.commonpath([working_dir_abs, target_dir]) == working_dir_abs

        if not valid_target_dir:
            raise Exception(f'Error: Cannot list "{directory}" as it is outside the permitted working directory')

        dir_contents = os.listdir(target_dir)
        files_info = []
        for name in dir_contents:
            full_path = os.path.join(target_dir, name)
            size = os.path.getsize(full_path)
            is_dir = os.path.isdir(full_path)
            files_info.append(
                f"- {name}: file_size={size} bytes, is_dir={is_dir}"
            )

        return "\n".join(files_info)
    except OSError as e:
        # catch any unforeseen errors raised by the os library and return a string
        return f"Error: {e}"
    except Exception as e:
        return e


schema_get_files_info = types.FunctionDeclaration(
    name="get_files_info",
    description="Lists files in a specified directory relative to the working directory, providing file size and directory status",
    parameters=types.Schema(
        type=types.Type.OBJECT,
        properties={
            "directory": types.Schema(
                type=types.Type.STRING,
                description="Directory path to list files from, relative to the working directory (default is the working directory itself)",
            ),
        },
    ),
)
